Refactor build file
